+++ src/prompt.js
@@ -1,5 +1,0 @@
-const { contextBridge, ipcRenderer } = require('electron');
-
-contextBridge.exposeInMainWorld('ipcRenderer', {
-  send: (channel, payload) => ipcRenderer.send(channel, payload)
-});